refactor(city): extract guild food shop menu into its own module
